Add Order Service

+++ BookStore.DataAccess/Entities/Entities/AuthorInPrintingEdition.cs
@@ -7,9 +7,9 @@ namespace BookStore.DataAccess.Entities.Enums
 {
     public class AuthorInPrintingEdition : BaseEntity
     {
-        public long AuthorId { get; set; }
+        public int AuthorId { get; set; }
         //public Author Author { get; set; }
-        public long PrintingEditionId { get; set; }
+        public int PrintingEditionId { get; set; }
         //public PrintingEdition PrintingEdition { get; set; }
     }
 }
